+++ src/common/path-helper.js
