+++ src/text/locale.js
